+++ platformio/managers/lib.py
@@ -104,6 +104,7 @@ class LibraryManager(BasePkgManager):
                 "sam": "atmelsam",
                 "samd": "atmelsam",
                 "esp8266": "espressif8266",
+                "esp32": "espressif32",
                 "arc32": "intel_arc32"
             }
             for arch in manifest['architectures'].split(","):