+++ HISTORY.rst
@@ -11,6 +11,9 @@ PlatformIO 3.0
 * Integration with `Jenkins CI <http://docs.platformio.org/en/latest/ci/jenkins.html>`_
 * Depend on development platform using VSC URL (Git, Mercurial and Subversion)
   in `Project Configuration File "platformio.ini" <http://docs.platformio.org/en/latest/projectconf/section_env_general.html#platform>`__
+* Improved handling of library dependencies defined as VCS or SemVer in
+  `Project Configuration File "platformio.ini" <http://docs.platformio.org/en/latest/projectconf/section_env_general.html#platform>`__
+  (`issue #1155 <https://github.com/platformio/platformio-core/issues/1155>`_)
 * Fixed "Super-Quick (Mac / Linux)" installer script
   (`issue #1017 <https://github.com/platformio/platformio-core/issues/1017>`_)
 * Fixed issue with "IOError" in VSCode when processing a project